% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{assign_cell_main}
\alias{assign_cell_main}
\title{Function to assign cell types through iterations}
\usage{
assign_cell_main(
  CelestaObj,
  max_iteration = 10,
  cell_change_threshold = 0.01,
  min_diff = 0,
  min_probability = 0,
  high_marker_threshold_anchor = rep(0.7, length = 50),
  low_marker_threshold_anchor = rep(0.9, length = 50),
  high_marker_threshold_iteration = rep(0.5, length = 50),
  low_marker_threshold_iteration = rep(1, length = 50)
)
}
\description{
Function to assign cell types through iterations
}
