% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{AssignCellTypes}
\alias{AssignCellTypes}
\title{AssignCellTypes}
\usage{
AssignCellTypes(
  initial_pri_matrix,
  current_cell_prob,
  current_cell_type_assignment,
  marker_exp_prob,
  cell_type_num,
  unassigned_cells,
  round,
  high_marker_threshold,
  low_marker_threshold,
  min_difference = 0,
  min_prob = 0
)
}
\arguments{
\item{initial_pri_matrix}{user defined cell-type marker matrix for a
specific round}

\item{current_cell_prob}{the current cell probability
(number_cells x number_cell_type)}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{marker_exp_prob}{the marker expression probability for each cell}

\item{cell_type_num}{the cell types associated with the current round}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}

\item{round}{the current round}

\item{high_marker_threshold}{the upper threshold for each cell type}

\item{low_marker_threshold}{the lower threshold for each cell type}

\item{min_difference}{lower bound used to determine cells that meet the
threshold}

\item{min_prob}{lower bound used to determine cells that meet the threshold}
}
\value{
an updated current cell type assignment (number_cells x total_rounds)
with more cells assigned for the current round
}
\description{
Find the cell types based on the scores (anchor cell) or
probabilities (index cell)
}
