% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{build_sigmoid_function}
\alias{build_sigmoid_function}
\title{Build sigmoid function for calculation of expression probability}
\usage{
build_sigmoid_function(marker_exp_matrix, figure = FALSE)
}
\description{
Build sigmoid function for calculation of expression probability
}
