% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{BuildSigmoidFunction}
\alias{BuildSigmoidFunction}
\title{BuildSigmoidFunction}
\usage{
BuildSigmoidFunction(marker_exp_matrix, figure = FALSE)
}
\arguments{
\item{marker_exp_matrix}{transformed protein marker expression (or original
segmentation protein marker expression if transformation is not specified)}

\item{figure}{whether a figure should be generated or not}
}
\value{
the sigmoid function parameter, containing the \eqn{x_root} and slope
}
\description{
Builds the sigmoid function for the calculation of the
expression probability
}
