% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CalcMarkerActivationProbability}
\alias{CalcMarkerActivationProbability}
\title{CalcMarkerActivationProbability}
\usage{
CalcMarkerActivationProbability(marker_exp_matrix, figure = FALSE)
}
\arguments{
\item{marker_exp_matrix}{transformed protein marker expression (or original
segmentation
protein marker expression if transformation is not specified)}
}
\value{
the protein marker activation probability
}
\description{
Calculates the activation probability for each marker in the
prior matrix
}
