% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CalculateBeta}
\alias{CalculateBeta}
\title{CalculateBeta}
\usage{
CalculateBeta(
  dist_from_nearest_assigned_cell,
  scale_factor = 5,
  bandwidth = 100
)
}
\arguments{
\item{dist_from_nearest_assigned_cell}{the distance from the nearest assigned
cell}

\item{scale_factor}{the scale factor}

\item{bandwidth}{the bandwidth}
}
\value{
the beta value
}
\description{
Calculates beta
}
