% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CalculateIndexCellProb}
\alias{CalculateIndexCellProb}
\title{CalculateIndexCellProb}
\usage{
CalculateIndexCellProb(
  current_cell_prob,
  current_cell_type_assignment,
  current_beta,
  nb_cell_type,
  current_scoring_matrix,
  cell_type_num,
  unassigned_cells,
  round
)
}
\arguments{
\item{current_cell_prob}{the current cell probability
(number_cells x number_cell_type)}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{current_beta}{the current beta values}

\item{nb_cell_type}{cell types of the neighboring cells for index cells}

\item{current_scoring_matrix}{the current scoring matrix
(number_cells x number_cell_type)}

\item{cell_type_num}{the cell types associated with the current round}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}

\item{round}{the current round}
}
\value{
calculates the probability for each cell type for unassigned cells
}
\description{
Calculates the probability for index cells
}
