% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CalculateProbabilityDifference}
\alias{CalculateProbabilityDifference}
\title{CalculateProbabilityDifference}
\usage{
CalculateProbabilityDifference(
  max.prob,
  max.prob_index,
  cell_prob_list,
  unassigned_cells
)
}
\arguments{
\item{max.prob}{the maximum marker probability for each cell}

\item{max.prob_index}{the index of the maximum marker probability for each
cell}

\item{cell_prob_list}{the probabilities of the cells are are not assigned a
cell type}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}
}
\value{
the minimum of the difference in probability between the maximum
marker probability and other marker probabilities
}
\description{
Calculate the probability differences
}
