% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CalculateScores}
\alias{CalculateScores}
\title{CalculateScores}
\usage{
CalculateScores(
  marker_exp_prob,
  current_pri_matrix,
  current_scoring_matrix,
  round,
  unassigned_cells,
  cell_type_num
)
}
\arguments{
\item{marker_exp_prob}{the marker expression probability for each cell}

\item{current_pri_matrix}{the updated cell-type marker matrix}

\item{current_scoring_matrix}{the current scoring matrix
(number_cells x number_cell_type)}

\item{round}{the current round}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}

\item{cell_type_num}{the cell types associated with the current round}
}
\value{
the current scoring matrix containing the scores for each cell type
associated with the current round for each unassigned cell
}
\description{
Calculate the scores based on the scoring function
}
