% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{cell_filtering}
\alias{cell_filtering}
\title{Filter cells that could potentially be artifacts}
\usage{
cell_filtering(
  high_marker_threshold = 0.9,
  low_marker_threshold = 0.4,
  CelestaObj
)
}
\description{
Filter cells that could potentially be artifacts
}
