% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{cell_prob}
\alias{cell_prob}
\title{Function to calculate probability for index cells}
\usage{
cell_prob(CelestaObj, cell_type_num, cells_to_check, round)
}
\description{
Function to calculate probability for index cells
}
