% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{cell_type}
\alias{cell_type}
\title{Find the cell types based on the scores (anchor cell) or probabilities (index cell)}
\usage{
cell_type(
  CelestaObj,
  cell_type_num,
  cells_to_check,
  round,
  min_difference = 0,
  min_prob = 0,
  high_marker_threshold,
  low_marker_threshold
)
}
\description{
Find the cell types based on the scores (anchor cell) or probabilities (index cell)
}
