% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CountCellType}
\alias{CountCellType}
\title{CountCellType}
\usage{
CountCellType(prior_info, current_cell_type_assignment, cell_type_num, round)
}
\arguments{
\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{cell_type_num}{the cell types associated with the current round}

\item{round}{the current round}
}
\value{
the count and proportion for each cell type based on the current cell
type assignments
}
\description{
Counts the cell type
}
