% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CreateCELESTAobj}
\alias{CreateCELESTAobj}
\title{Create CELESTA object}
\usage{
CreateCELESTAobj(
  project_title = "Project",
  prior_marker_info,
  imaging_data_file,
  cofactor = 10,
  transform_type = 1,
  number_of_neighbors = 5,
  bandwidth = 100
)
}
\description{
Create CELESTA object
}
