% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{CreateCelestaObject}
\alias{CreateCelestaObject}
\title{CreateCelestaObject}
\usage{
CreateCelestaObject(
  project_title,
  prior_marker_info,
  imaging_data_file,
  cofactor = 10,
  transform_type = 1,
  number_of_neighbors = 5,
  bandwidth = 100
)
}
\arguments{
\item{project_title}{\emph{required} name of the project (used in file names)}

\item{prior_marker_info}{\emph{required} user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{imaging_data_file}{\emph{required} segmented imaging data.
The first column must contain the cell types to be inferred. The second
column must contain the lineage information with the following format
(without spaces): # _ # _ #.
\itemize{
\item The first number indicates round. Cell types with the same lineage level
are inferred at the same round. An higher number indicates higher cell-type
resolution. For example, immune cells -> CD3+ T cells -> CD4+ T cells.
\item The second number indicates the previous lineage cell type number for the
current cell type. For example, the second number for CD3+ T cell is 5
because it is a subtype of immune cells which have cell type number 5.
\item The third number is a number assigned to the cell type
(i.e. cell type number).
}

The third column and beyond are columns for protein markers.
\itemize{
\item If a protein marker is known to be expressed for that cell type, then it
is denoted by a "1".
\item If a protein marker is known to not express for a cell type, then it is
denoted by a "0".
\item If the protein marker is irrelevant or uncertain to express for a cell
type, then it is denoted by "NA".
}}

\item{cofactor}{value used to calculate the arcsinh transform on the protein
marker expressions}

\item{transform_type}{indicates a transform type for the protein marker
expressions (0 = no transform, 1 = arcsinh transform)}

\item{number_of_neighbors}{the number of cells in a single neighborhood}

\item{bandwidth}{the upper distance bound used when calculating
neighborhoods by distance}
}
\value{
an initialized Celesta object
}
\description{
Initializes the following fields of the Celesta object:
\itemize{
\item \code{cell_ID}
\item \code{original_exp}
\item \code{marker_exp_matrix}
\item \code{prior_info}
\item \code{lineage_info}
\item \code{total_rounds}
\item \code{coords}
\item \code{marker_exp_prob}
\item \code{nb_list}
\item \code{cell_nb_in_bandwidth}
\item \code{cell_nb_dist}
\item \code{current_cell_type_assignment}
\item \code{anchor_cell_type_assignment}
\item \code{starting_cell_type_assignment}
\item \code{current_scoring_matrix}
\item \code{current_cell_prob}
}
}
