% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{FilterArtifactCells}
\alias{FilterArtifactCells}
\title{FilterArtifactCells}
\usage{
FilterArtifactCells(
  total_rounds,
  marker_exp_matrix,
  marker_exp_prob,
  current_cell_type_assignment,
  high_marker_threshold = 0.9,
  low_marker_threshold = 0.4
)
}
\arguments{
\item{total_rounds}{the maximum round value}

\item{marker_exp_matrix}{transformed protein marker expression (or original
segmentation protein marker expression if transformation is not specified)}

\item{marker_exp_prob}{the marker expression probability for each cell}

\item{current_cell_type_assignment}{the cell type assignments for each round
for each cell}

\item{high_marker_threshold}{upper bound used to filter out questionable
cells}

\item{low_marker_threshold}{lower bound used to filter out questionable
cells}
}
\value{
current cell type assignment, where a questionable cells are marked
with a row of NAs.
}
\description{
Filter out cells that could potentially be artifacts
}
