% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{FilterCells}
\alias{FilterCells}
\title{FilterCells}
\usage{
FilterCells(
  celesta_obj,
  high_marker_threshold = 0.9,
  low_marker_threshold = 0.4
)
}
\arguments{
\item{celesta_obj}{an initialized Celesta object (provided by
\code{CreateCelestaObject})}

\item{high_marker_threshold}{upper bound used to filter out questionable
cells}

\item{low_marker_threshold}{lower bound used to filter out questionable
cells}
}
\value{
a Celesta object with questionable cells marked with NA
}
\description{
Filters out artifact cells from the cell type assignments
}
