% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{find_min_prob_diff}
\alias{find_min_prob_diff}
\title{Calculate probability differences}
\usage{
find_min_prob_diff(max.prob, max.prob_index, cell_prob_list, cells_to_check)
}
\description{
Calculate probability differences
}
