% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{FindCellsWithId}
\alias{FindCellsWithId}
\title{FindCellsWithId}
\usage{
FindCellsWithId(current_cell_type_assignment, lineage_info, cell_ID, round)
}
\arguments{
\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{lineage_info}{the lineage information from \code{prior_info} parsed into
round, previous cell type, and cell type number columns}

\item{cell_ID}{the IDs of the cells (from 1 to the total number of cells)}

\item{round}{the current round}
}
\value{
cells that have been assigned a cell type
}
\description{
Find cells assigned with a cell type
}
