% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{FitGmmModel}
\alias{FitGmmModel}
\title{FitGmmModel}
\usage{
FitGmmModel(marker_exp, marker_name, figure = FALSE)
}
\arguments{
\item{marker_exp}{the expression of the marker for each cell}

\item{marker_name}{the name of the marker}

\item{figure}{whether a figure should be generated or not}
}
\value{
the Gaussian mixture model parameters for the marker
}
\description{
Fits a Gaussian mixture model for each marker
}
