% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{get_dist_from_nearest_assigned_cells}
\alias{get_dist_from_nearest_assigned_cells}
\title{Get distance from nearest assigned cells}
\usage{
get_dist_from_nearest_assigned_cells(
  CelestaObj,
  cell_type_num,
  cells_to_check,
  cells_with_ID,
  round
)
}
\description{
Get distance from nearest assigned cells
}
