% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{get_final_inferred_cell_types}
\alias{get_final_inferred_cell_types}
\title{Get final results}
\usage{
get_final_inferred_cell_types(total_rounds, CelestaObj, imaging_data)
}
\description{
Get final results
}
