% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{get_initial_prior_matrix}
\alias{get_initial_prior_matrix}
\title{For each round, need to get the prior knowledge on the cell types}
\usage{
get_initial_prior_matrix(CelestaObj, round)
}
\description{
For each round, need to get the prior knowledge on the cell types
}
