% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetDistFromNearestAssignedCells}
\alias{GetDistFromNearestAssignedCells}
\title{GetDistFromNearestAssignedCells}
\usage{
GetDistFromNearestAssignedCells(
  cell_nb_in_bandwidth,
  cell_nb_dist,
  current_cell_type_assignment,
  cell_type_num,
  unassigned_cells,
  assigned_cells,
  round
)
}
\arguments{
\item{cell_nb_in_bandwidth}{the cells located within a bandwidth to cell \emph{c}}

\item{cell_nb_dist}{the distance of each cell to cell \emph{c} within a bandwidth}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{cell_type_num}{the cell types associated with the current round}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}

\item{assigned_cells}{cells with an assigned cell type}

\item{round}{the current round}
}
\value{
the distance to the nearest assigned cells
}
\description{
Get distance from nearest assigned cells
}
