% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetFinalInferredCellTypes}
\alias{GetFinalInferredCellTypes}
\title{GetFinalInferredCellTypes}
\usage{
GetFinalInferredCellTypes(
  project_name,
  total_rounds,
  current_cell_type_assignment,
  anchor_cell_type_assignment,
  prior_info,
  lineage_info,
  coords,
  original_exp,
  save_result = T
)
}
\arguments{
\item{total_rounds}{the maximum round}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{anchor_cell_type_assignment}{the anchor cell type assignments}

\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{lineage_info}{the lineage information from \code{prior_info} parsed into
round, previous cell type, and cell type number columns}

\item{coords}{the x, y coordinates of each cell}

\item{original_exp}{original protein marker expression (containing only the
protein markers specified in \code{prior_info})}

\item{save_data}{whether or not to save the final cell type assignment
and anchor cell assignment results}
}
\value{
the final cell type assignments
}
\description{
Get final cell types and writes two files: the final cell type
assignments and the anchor cell type assignments.
}
