% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetInitialPriorMatrix}
\alias{GetInitialPriorMatrix}
\title{GetInitialPriorMatrix}
\usage{
GetInitialPriorMatrix(lineage_info, prior_marker_info, round)
}
\arguments{
\item{lineage_info}{the lineage information from \code{prior_info} parsed into
round, previous cell type, and cell type number columns}

\item{round}{the current round}
}
\value{
the prior knowledge of the cells types with the specified round.
}
\description{
Gets the prior knowledge of the cell types with the specified
round.
}
