% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetMarkerExpMatrix}
\alias{GetMarkerExpMatrix}
\title{Get protein marker expression}
\usage{
GetMarkerExpMatrix(
  CelestaObj,
  prior_marker_info,
  imaging_data_file,
  cofactor,
  transform_type
)
}
\description{
Get protein marker expression
}
