% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetPriorInfo}
\alias{GetPriorInfo}
\title{GetPriorInfo}
\usage{
GetPriorInfo(prior_marker_info)
}
\arguments{
\item{prior_marker_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}
}
\value{
a list with the following information:
\describe{
\item{\code{lineage_info}}{the lineage information parsed into round, previous
cell type, and cell type number columns}
\item{\code{total_rounds}}{the maximum round value}
}
}
\description{
Extracts the lineage information from  the \code{prior_marker_info}
and determines the total rounds
}
