% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetScore}
\alias{GetScore}
\title{GetScore}
\usage{
GetScore(activation_prob_to_use, prior_info, non_NA_index)
}
\arguments{
\item{activation_prob_to_use}{the marker expression probabilities of the
unassigned cells}

\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{non_NA_index}{the index of the columns in \code{current_pri_matrix} that do
not contain NA for a particular cell}
}
\value{
the score of the cell
}
\description{
Calculate scores using MSE
}
