% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GMM_fitting}
\alias{GMM_fitting}
\title{Gaussian mixture model for each marker}
\usage{
GMM_fitting(marker_exp, marker_name, figure = FALSE)
}
\description{
Gaussian mixture model for each marker
}
