% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{InitializeCellAndScoringMatrices}
\alias{InitializeCellAndScoringMatrices}
\title{InitializeCellAndScoringMatrices}
\usage{
InitializeCellAndScoringMatrices(lineage_info, marker_exp_matrix, prior_info)
}
\arguments{
\item{lineage_info}{the lineage information from \code{prior_info} parsed into
round, previous cell type, and cell type number columns}

\item{marker_exp_matrix}{transformed protein marker expression (or original
segmentation protein marker expression if transformation is not specified)}

\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}
}
\value{
a list with the following information
\describe{
\item{\code{current_cell_type_assignment}}{a zero matrix with dimension
(number_cells x total_rounds)}
\item{\code{current_scoring_matrix}}{a NA matrix with dimension
(number_cells x number_cell_type)}
\item{\code{current_cell_prob}}{a NA matrix with dimension
(number_cells x number_cell_type)}
}
}
\description{
Initialize the cell type assignments, cell probabilities, and
scoring matrices
}
