% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{marker_activation_probability}
\alias{marker_activation_probability}
\title{Calculate expression probability for each marker in the prior matrix}
\usage{
marker_activation_probability(CelestaObj, figure = FALSE)
}
\description{
Calculate expression probability for each marker in the prior matrix
}
