% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{MarkQuestionableCells}
\alias{MarkQuestionableCells}
\title{Mark questionable cells}
\usage{
MarkQuestionableCells(
  cell_activation_prob,
  high_marker_threshold,
  low_marker_threshold
)
}
\description{
Mark questionable cells
}
