% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{neighbor_cell_type}
\alias{neighbor_cell_type}
\title{Find the cell types of the neighbors for cells_to_check}
\usage{
neighbor_cell_type(CelestaObj, cell_type_num, round, cells_to_check)
}
\description{
Find the cell types of the neighbors for cells_to_check
}
