% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{NeighborCellType}
\alias{NeighborCellType}
\title{NeighborCellType}
\usage{
NeighborCellType(
  nb_list,
  current_cell_type_assignment,
  cell_type_num,
  round,
  unassigned_cells
)
}
\arguments{
\item{nb_list}{the list of N-nearest neighbors}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{cell_type_num}{the cell types associated with the current round}

\item{round}{the current round}

\item{unassigned_cells}{cells not assigned a cell type for each round and
iteration}
}
\value{
the cell types of the neighbors of unassigned cells
}
\description{
Find the cell types of the neighbors of unassigned cells
}
