% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{plot_cells_any_combination}
\alias{plot_cells_any_combination}
\title{Plot the cells using XY coordinates}
\usage{
plot_cells_any_combination(
  cell_type_assignment_to_plot,
  CelestaObj,
  cell_number_to_use,
  cell_type_colors = c(palette()[2:7], "white"),
  test_size = 1
)
}
\description{
Plot the cells using XY coordinates
}
