% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{plot_exp_prob}
\alias{plot_exp_prob}
\title{Plot the expression probabilities of cells in the tissue}
\usage{
plot_exp_prob(CelestaObj, size_to_use = 1, width_to_use = 5, height_to_use = 4)
}
\description{
Plot the expression probabilities of cells in the tissue
}
