% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{PlotCellsAnyCombination}
\alias{PlotCellsAnyCombination}
\title{PlotCellsAnyCombination}
\usage{
PlotCellsAnyCombination(
  cell_type_assignment_to_plot,
  coords,
  prior_info,
  cell_number_to_use,
  cell_type_colors = c(palette()[2:7], "white"),
  test_size = 1,
  save_plot = TRUE,
  output_dir = "."
)
}
\arguments{
\item{cell_type_assignment_to_plot}{the final cell type assignment for each
cell}

\item{coords}{the x, y coordinates of each cell}

\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{cell_number_to_use}{the row number of the cell types to plot from
\code{prior_info}. To plot unknown cells, include 0 in the list.}

\item{cell_type_colors}{the colors for the cell types. If unknown cells
are plotted, the color of the unknown cells will be the last color listed.}

\item{test_size}{the size of the points in the plot}

\item{save_plot}{whether to save the plot}

\item{output_dir}{the path to the directory to where the plot will be
outputted. This defaults to the directory containing CELESTA_functions.R.
Note that the directory must exist.}
}
\value{
writes the final cell type assignment plot
}
\description{
Plots the cells using x, y coordinates with their assigned cell
types
}
