% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{PlotExpProb}
\alias{PlotExpProb}
\title{PlotExpProb}
\usage{
PlotExpProb(
  celesta_obj,
  marker_to_use,
  size_to_use = 1,
  width_to_use = 5,
  height_to_use = 4
)
}
\arguments{
\item{celesta_obj}{an fully initialized Celesta object (provided by
\code{AssignCells})}

\item{size_to_use}{the size of the points in the plot}

\item{width_to_use}{the width of the plot}

\item{height_to_use}{the height of the plot}
}
\value{
writes a plot of the expression probabilities for each marker
}
\description{
Plots the expression probabilities of cells in the tissue
}
