% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{PlotSingleExpProb}
\alias{PlotSingleExpProb}
\title{PlotSingleExpProb}
\usage{
PlotSingleExpProb(
  coords,
  marker_exp_prob,
  cols = NULL,
  marker_to_use,
  size_to_use = 1,
  width_to_use = 5,
  height_to_use = 4,
  save_plot = TRUE,
  output_dir = "."
)
}
\arguments{
\item{coords}{the x, y coordinates of each cell}

\item{marker_exp_prob}{the marker expression probability for each cell}

\item{cols}{the color palette for the plot}

\item{marker_to_use}{marker to plot}

\item{size_to_use}{the size of the points in the plot}

\item{width_to_use}{the width of the plot}

\item{height_to_use}{the height of the plot}

\item{save_plot}{whether to save the plot}

\item{output_dir}{the path to the directory to where the plot will be
outputted. This defaults to the directory containing CELESTA_functions.R.
Note that the directory must exist.}
}
\value{
generates a plot of the expression probabilities for a specified
marker
}
\description{
Plots the expression probabilities of cells in the tissue. This
is use soley for the Shiny app.
}
