% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{scoring_function}
\alias{scoring_function}
\title{Function for calculating scoring function}
\usage{
scoring_function(CelestaObj, round, cells_to_check, cell_type_num)
}
\description{
Function for calculating scoring function
}
