% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{update_prior_matrix}
\alias{update_prior_matrix}
\title{Function to update prior knowledge matrix of the cell type signatures}
\usage{
update_prior_matrix(CelestaObj, round, cell_type_num)
}
\description{
Function to update prior knowledge matrix of the cell type signatures
}
