% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{UpdatePriorMatrix}
\alias{UpdatePriorMatrix}
\title{UpdatePriorMatrix}
\usage{
UpdatePriorMatrix(
  current_pri_matrix,
  initial_pri_matrix,
  current_cell_type_assignment,
  marker_exp_prob,
  round,
  cell_type_num
)
}
\arguments{
\item{current_pri_matrix}{the updated cell-type marker matrix}

\item{initial_pri_matrix}{user defined cell-type marker matrix for a specific
round}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{marker_exp_prob}{the marker expression probability for each cell}

\item{round}{the current round}

\item{cell_type_num}{the cell types associated with the current round}
}
\value{
updates the prior knowledge matrix with information from cells
assigned to each cell type
}
\description{
Updates prior knowledge matrix of the cell type signatures
}
