% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetFinalInferredCellTypes}
\alias{GetFinalInferredCellTypes}
\title{GetFinalInferredCellTypes}
\usage{
GetFinalInferredCellTypes(
  project_name,
  total_rounds,
  current_cell_type_assignment,
  anchor_cell_type_assignment,
  prior_info,
  lineage_info,
  imaging_data
)
}
\arguments{
\item{total_rounds}{the maximum round}

\item{current_cell_type_assignment}{the current cell type assignments
(number_cells x total_rounds)}

\item{anchor_cell_type_assignment}{the anchor cell type assignments}

\item{prior_info}{user-defined cell-type signature matrix.

The data should contain two columns (name X and Y) for the x, y coordinates
and a column for each protein marker. Each row represents the data for a
single cell, including its x, y coordinates and expression for each protein
marker.}

\item{lineage_info}{the lineage information from \code{prior_info} parsed into
round, previous cell type, and cell type number columns}

\item{imaging_data}{segmented imaging data.
The first column must contain the cell types to be inferred. The second
column must contain the lineage information with the following format
(without spaces): # _ # _ #.
\itemize{
\item The first number indicates round. Cell types with the same lineage level
are inferred at the same round. An higher number indicates higher cell-type
resolution. For example, immune cells -> CD3+ T cells -> CD4+ T cells.
\item The second number indicates the previous lineage cell type number for the
current cell type. For example, the second number for CD3+ T cell is 5
because it is avsubtype of immune cells which have cell type number 5.
\item The third number is a number assigned to the cell type
(i.e. cell type number).
}

The third column and beyond are columns for protein markers.
\itemize{
\item If a protein marker is known to be expressed for that cell type, then it
is denoted by a "1".
\item If a protein marker is known to not express for a cell type, then it is
denoted by a "0".
\item If the protein marker is irrelevant or uncertain to express for a cell
type, then it is denoted by "NA".
}}
}
\value{
the final cell type assignments
}
\description{
Get final cell types and writes two files: the final cell type
assignments and the anchor cell type assignments.
}
