% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_functions.R
\name{GetNeighborInfo}
\alias{GetNeighborInfo}
\title{Get neighborhood informtion}
\usage{
GetNeighborInfo(CelestaObj, number_of_neighbors = 5, bandwidth = 100)
}
\description{
Get neighborhood informtion
}
